Send notifications to all devices (see description)
To make this work you capture the token from the console. After that you use swagger to enter the token using RegisterToken endpoint. After which you do this for all devices you want to send the notification to. Then simply use the test to all button in the frontend sender application and voila.

Side note: strange thing was that for the physical device I first had to send a notification directly to that device before sending to the topic started working